// backend/src/models/StockQuote.ts
export interface StockQuote {
  symbol: string;
  shortName?: string;
  longName?: string;
  regularMarketPrice?: number;
  regularMarketChange?: number;
  regularMarketChangePercent?: number;
  regularMarketTime?: Date;
  currency?: string;
  exchange?: string;
  quoteType?: string;
  marketState?: string;
  marketCap?: number;
  volume?: number;
  averageVolume?: number;
  dayHigh?: number;
  dayLow?: number;
  fiftyTwoWeekHigh?: number;
  fiftyTwoWeekLow?: number;
}